add confirmation to d to deletion of question

+++ app/routes/question.js
@@ -8,15 +8,19 @@ export default Ember.Route.extend({
       newQuestion.save();
       this.transitionTo('/');
     },
+
     deleteQuestion(question) {
-      var answer_deletions = question.get('answers').map(function(answer) {
-        return answer.destroyRecord();
-      });
-      Ember.RSVP.all(answer_deletions).then(function() {
-        return question.destroyRecord();
-      })
-      this.transitionTo('index');
+      if(confirm('Are you sure you want to delete this question?')) {
+        var answer_deletions = question.get('answers').map(function(answer) {
+          return answer.destroyRecord();
+        });
+        Ember.RSVP.all(answer_deletions).then(function() {
+          return question.destroyRecord();
+        })
+        this.transitionTo('index');
+      }
     },
+
     updateQuestion(question, params) {
       Object.keys(params).forEach(function(key) {
         if(params[key]!==undefined) {
